initial typescript setup
