+++ package.json
@@ -2,11 +2,11 @@
   "name": "simple-news-api",
   "version": "1.0.0",
   "description": "Simple News API",
-  "main": "server.js",
+  "main": "bin/server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "dev": "nodemon --inspect server.js"
+    "start": "node bin/server.js",
+    "dev": "nodemon --inspect bin/server.js"
   },
   "repository": {
     "type": "git",
@@ -22,9 +22,15 @@
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.2",
+    "lodash": "^4.17.21",
     "puppeteer": "^13.0.1"
   },
   "devDependencies": {
-    "nodemon": "^2.0.15"
+    "@types/express": "^4.17.13",
+    "@types/lodash": "^4.14.178",
+    "@types/puppeteer": "^5.4.4",
+    "nodemon": "^2.0.15",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.5"
   }
 }
