+++ tsconfig.json
@@ -1,0 +1,24 @@
+/* Docs: https://www.typescriptlang.org/tsconfig */
+{
+  "compilerOptions": {
+    "target": "ES2020",
+    "module": "commonjs",
+    "outDir": "bin",
+    "rootDir": ".",
+    "strict": true,
+    "noImplicitAny": false,
+    "strictNullChecks": true,
+    "checkJs": true,
+    "allowJs": true,
+    "moduleResolution": "node",
+    "types": ["node"],
+    "lib": ["es6", "dom"],
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "allowUnreachableCode": true,
+    "useUnknownInCatchVariables": false // TypeScript 4.4+ only!
+  },
+  "include": ["./*.ts", "./*.js"],
+  "exclude": ["node_modules", "<node_internals>/**"]
+}
